Update README with getting started and add Postman collections - DELETE methods shall verify that ID exists and return 404 if not - Add Postman env variables and collections for companies and employees - Document how to build, run and test APIs in README

+++ web/src/main/java/no/acntech/project101/web/company/resources/CompanyResource.java
@@ -5,7 +5,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,7 +12,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -84,8 +82,14 @@ public class CompanyResource {
     }
 
     @DeleteMapping("{id}")
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    public void deleteCompany(@PathVariable final Long id) {
-        companyService.delete(id);
+    public ResponseEntity deleteCompany(@PathVariable final Long id) {
+        final Optional<Company> company = companyService.findById(id);
+
+        if (company.isPresent()) {
+            companyService.delete(id);
+            return ResponseEntity.accepted().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
